+++ src/main/java/site/praytogether/pray_together/domain/fcm_token/repository/FcmTokenRepository.java
@@ -1,10 +1,10 @@
-package site.praytogether.pray_together.domain.notification.repository;
+package site.praytogether.pray_together.domain.fcm_token.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import site.praytogether.pray_together.domain.fcm_token.model.FcmToken;
 import site.praytogether.pray_together.domain.member.model.Member;
-import site.praytogether.pray_together.domain.notification.model.FcmToken;
 
 public interface FcmTokenRepository extends JpaRepository<FcmToken, Long> {
   boolean existsByTokenAndMember(String token, Member member);
