--- src/main/java/site/praytogether/pray_together/domain/notification/event/EventPublisher.java
@@ -1,5 +1,0 @@
-package site.praytogether.pray_together.domain.notification.event;
-
-public interface EventPublisher {
-  public void publishPrayerComplete();
-}
