--- src/main/java/site/praytogether/pray_together/domain/notification/event/NotificationEventPublisher.java
@@ -1,10 +1,0 @@
-package site.praytogether.pray_together.domain.notification.event;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class NotificationEventPublisher implements EventPublisher {
-
-  @Override
-  public void publishPrayerComplete() {}
-}
