+++ src/main/java/site/praytogether/pray_together/domain/prayer/application/PrayerApplicationService.java
@@ -7,12 +7,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.praytogether.pray_together.domain.base.MessageResponse;
+import site.praytogether.pray_together.domain.fcm_token.model.FcmToken;
+import site.praytogether.pray_together.domain.fcm_token.service.FcmTokenService;
 import site.praytogether.pray_together.domain.member.model.Member;
 import site.praytogether.pray_together.domain.member.service.MemberService;
 import site.praytogether.pray_together.domain.member_room.service.MemberRoomService;
-import site.praytogether.pray_together.domain.notification.model.FcmToken;
-import site.praytogether.pray_together.domain.notification.service.FcmTokenService;
-import site.praytogether.pray_together.domain.notification.service.NotificationGateway;
+import site.praytogether.pray_together.domain.notification.gateway.NotificationGateway;
 import site.praytogether.pray_together.domain.notification.service.PrayerCompletionNotificationService;
 import site.praytogether.pray_together.domain.prayer.dto.PrayerCompletionCreateRequest;
 import site.praytogether.pray_together.domain.prayer.dto.PrayerContentResponse;
